bloom

Module bloom

Structs

BloomFilter

A standard BloomFilter. If an item is instered then contains is guaranteed to return true for that item. For items not inserted contains will probably return false. The probability that contains returns true for an item that was not inserted is called the False Positive Rate.

Functions

needed_bits

Return the number of bits needed to satisfy the specified false positive rate, if the filter will hold num_items items.

optimal_num_hashes

Return the optimal number of hashes to use for the given number of bits and items in a filter