Struct bloom::BloomFilter

pub struct BloomFilter { /* fields omitted */ }

Space-efficient probabilistic hash set

Methods

impl BloomFilter

fn new_with_fp(n_elems: usize, fp_rate: f32) -> BloomFilter

Build a Bloom Filter with a specified false positive rate

Arguments

n_elems: expected number of elements

fp_rate: desired false positive rate (0.0 -> 1.0)

fn new_with_size(n_elems: usize, size: usize) -> BloomFilter

Create a new Bloom Filter with specified buffer size

Arguments

n_elems: expected number of elements

size: desired buffer size

fn insert<T>(&mut self, e: &T) where
T: Hash,

Insert a member

Arguments

e: element to add

fn may_contain<T>(&self, e: &T) -> bool where
T: Hash,

Check membership

Arguments

e: element to check membership of

fn size(&self) -> usize

Number of elements in the BloomFilter

fn buckets(&self) -> usize

Number of buckets that a memebr can occupy

fn n_hashers(&self) -> usize

Number of hashers being used

fn fp_rate(&self) -> f32

False positive rate