bloom

Trait Intersectable

pub trait Intersectable {
    // Required method
    fn intersect(&mut self, other: &Self) -> bool;
}

Filters that implement this trait can be intersected with filters of the same type to produce a filter that contains the items that have been inserted into both filters.

Both filters MUST be the same size and be using the same hash functions for this to work. Will panic if the filters are not the same size, but will simply produce incorrect (meaningless) results if the filters are using different hash functions.

Required Methods

fn intersect(&mut self, other: &Self) -> bool

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.